tests/Element-getAttribute.html

<!DOCTYPE html>
<title>Element.getAttribute</title>
<script src="testharness.js"></script>
<div id="log"></div>
<script>
test(function() {
  var el = document.createElement("foo")
  assert_equals(el.getAttribute("bar"), null)

  el.setAttributeNS("ab", "attr", "pass")
  el.setAttributeNS("kl", "attr", "fail")
  assert_equals(el.getAttribute("attr"), "pass")

  el.setAttribute("style", "color:#fff;")
  assert_equals(el.getAttribute("style"), "color:#fff;")

  el = document.createElement("div")
  el.setAttributeNS("", "ALIGN", "left")
  assert_equals(el.getAttribute("ALIGN"), null)
  assert_equals(el.getAttribute("align"), null)

  el.setAttributeNS("", "CHEEseCaKe", "tasty")
  assert_equals(el.getAttribute("CHEESECAKE"), null)
  assert_equals(el.getAttribute("CHEEseCaKe"), null)
  assert_equals(el.getAttribute("cheesecake"), null)

  el = document.createElement("div")
  el.setAttributeNS(null, "align", "left")
  el.setAttributeNS("xx", "align", "right")
  el.setAttributeNS(null, "foo", "left")
  el.setAttributeNS("xx", "foo", "right")
  document.body.appendChild(el)
  assert_equals(el.getAttribute("align"), "left")
  assert_equals(el.getAttribute("foo"), "left")

  el = document.createElement("div")
  el.setAttributeNS("xx", "align", "right")
  el.setAttributeNS(null, "align", "left")
  el.setAttributeNS("xx", "foo", "right")
  el.setAttributeNS(null, "foo", "left")
  document.body.appendChild(el)
  assert_equals(el.getAttribute("align"), "right")
  assert_equals(el.getAttribute("foo"), "right")
})
</script>
